Restructured code to OOP. Please see details.
Removed code from Main class and created Engine, RunCalc and ConsoleMsgs classes. RunCalc calls all the functions which is then encapsulated in a single function called RunCalculator. In Engine class, I initially had trouble calling the variables of num1 and num2. Utilising a constructor to pass user input into the engine class. I also had to define the parameters in the Engine class to num1, num2 and the operator.